refactor job validators
