+++ core/domain/job_validators.py
@@ -1,0 +1,125 @@
+# coding: utf-8
+#
+# Copyright 2020 The Oppia Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS-IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Validators for job models."""
+
+from __future__ import absolute_import  # pylint: disable=import-only-modules
+from __future__ import unicode_literals  # pylint: disable=import-only-modules
+
+from core.domain import base_model_validators
+from core.platform import models
+import utils
+
+(
+    job_models,
+) = models.Registry.import_models([
+    models.NAMES.job
+])
+
+
+class JobModelValidator(base_model_validators.BaseModelValidator):
+    """Class for validating JobModels."""
+
+    @classmethod
+    def _get_model_id_regex(cls, item):
+        # Valid id: [job_type]-[current time]-[random int]
+        regex_string = '^%s-\\d*-\\d*$' % item.job_type
+        return regex_string
+
+    @classmethod
+    def _get_external_id_relationships(cls, item):
+        return []
+
+    @classmethod
+    def _validate_time_fields(cls, item):
+        """Validate the time fields in entity.
+
+        Args:
+            item: ndb.Model. JobModel to validate.
+        """
+        if item.time_started_msec and (
+                item.time_queued_msec > item.time_started_msec):
+            cls._add_error(
+                'time queued check',
+                'Entity id %s: time queued %s is greater '
+                'than time started %s' % (
+                    item.id, item.time_queued_msec, item.time_started_msec))
+
+        if item.time_finished_msec and (
+                item.time_started_msec > item.time_finished_msec):
+            cls._add_error(
+                'time started check',
+                'Entity id %s: time started %s is greater '
+                'than time finished %s' % (
+                    item.id, item.time_started_msec, item.time_finished_msec))
+
+        current_time_msec = utils.get_current_time_in_millisecs()
+        if item.time_finished_msec > current_time_msec:
+            cls._add_error(
+                'time finished check',
+                'Entity id %s: time finished %s is greater '
+                'than the current time' % (
+                    item.id, item.time_finished_msec))
+
+    @classmethod
+    def _validate_error(cls, item):
+        """Validate error is not None only if status is not canceled
+        or failed.
+
+        Args:
+            item: ndb.Model. JobModel to validate.
+        """
+        if item.error and item.status_code not in [
+                job_models.STATUS_CODE_FAILED, job_models.STATUS_CODE_CANCELED]:
+            cls._add_error(
+                base_model_validators.ERROR_CATEGORY_ERROR_CHECK,
+                'Entity id %s: error: %s for job is not empty but '
+                'job status is %s' % (item.id, item.error, item.status_code))
+
+        if not item.error and item.status_code in [
+                job_models.STATUS_CODE_FAILED, job_models.STATUS_CODE_CANCELED]:
+            cls._add_error(
+                base_model_validators.ERROR_CATEGORY_ERROR_CHECK,
+                'Entity id %s: error for job is empty but '
+                'job status is %s' % (item.id, item.status_code))
+
+    @classmethod
+    def _validate_output(cls, item):
+        """Validate output for entity is present only if status is
+        completed.
+
+        Args:
+            item: ndb.Model. JobModel to validate.
+        """
+        if item.output and item.status_code != job_models.STATUS_CODE_COMPLETED:
+            cls._add_error(
+                base_model_validators.ERROR_CATEGORY_OUTPUT_CHECK,
+                'Entity id %s: output: %s for job is not empty but '
+                'job status is %s' % (item.id, item.output, item.status_code))
+
+        if item.output is None and (
+                item.status_code == job_models.STATUS_CODE_COMPLETED):
+            cls._add_error(
+                base_model_validators.ERROR_CATEGORY_OUTPUT_CHECK,
+                'Entity id %s: output for job is empty but '
+                'job status is %s' % (item.id, item.status_code))
+
+    @classmethod
+    def _get_custom_validation_functions(cls):
+        return [
+            cls._validate_time_fields,
+            cls._validate_error,
+            cls._validate_output]
