+++ core/domain/prod_validators.py
@@ -65,14 +65,14 @@ import utils
 (
     base_models, collection_models, config_models,
     email_models, exp_models, feedback_models,
-    improvements_models, job_models, question_models,
+    improvements_models, question_models,
     recommendations_models, skill_models, stats_models,
     story_models, subtopic_models, suggestion_models,
     topic_models, user_models
 ) = models.Registry.import_models([
     models.NAMES.base_model, models.NAMES.collection, models.NAMES.config,
     models.NAMES.email, models.NAMES.exploration, models.NAMES.feedback,
-    models.NAMES.improvements, models.NAMES.job, models.NAMES.question,
+    models.NAMES.improvements, models.NAMES.question,
     models.NAMES.recommendations, models.NAMES.skill, models.NAMES.statistics,
     models.NAMES.story, models.NAMES.subtopic, models.NAMES.suggestion,
     models.NAMES.topic, models.NAMES.user
@@ -1680,101 +1680,6 @@ class UnsentFeedbackEmailModelValidator(
     @classmethod
     def _get_custom_validation_functions(cls):
         return [cls._validate_entity_type_and_entity_id_feedback_reference]
-
-
-class JobModelValidator(base_model_validators.BaseModelValidator):
-    """Class for validating JobModels."""
-
-    @classmethod
-    def _get_model_id_regex(cls, item):
-        # Valid id: [job_type]-[current time]-[random int]
-        regex_string = '^%s-\\d*-\\d*$' % item.job_type
-        return regex_string
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        return []
-
-    @classmethod
-    def _validate_time_fields(cls, item):
-        """Validate the time fields in entity.
-
-        Args:
-            item: ndb.Model. JobModel to validate.
-        """
-        if item.time_started_msec and (
-                item.time_queued_msec > item.time_started_msec):
-            cls._add_error(
-                'time queued check',
-                'Entity id %s: time queued %s is greater '
-                'than time started %s' % (
-                    item.id, item.time_queued_msec, item.time_started_msec))
-
-        if item.time_finished_msec and (
-                item.time_started_msec > item.time_finished_msec):
-            cls._add_error(
-                'time started check',
-                'Entity id %s: time started %s is greater '
-                'than time finished %s' % (
-                    item.id, item.time_started_msec, item.time_finished_msec))
-
-        current_time_msec = utils.get_current_time_in_millisecs()
-        if item.time_finished_msec > current_time_msec:
-            cls._add_error(
-                'time finished check',
-                'Entity id %s: time finished %s is greater '
-                'than the current time' % (
-                    item.id, item.time_finished_msec))
-
-    @classmethod
-    def _validate_error(cls, item):
-        """Validate error is not None only if status is not canceled
-        or failed.
-
-        Args:
-            item: ndb.Model. JobModel to validate.
-        """
-        if item.error and item.status_code not in [
-                job_models.STATUS_CODE_FAILED, job_models.STATUS_CODE_CANCELED]:
-            cls._add_error(
-                base_model_validators.ERROR_CATEGORY_ERROR_CHECK,
-                'Entity id %s: error: %s for job is not empty but '
-                'job status is %s' % (item.id, item.error, item.status_code))
-
-        if not item.error and item.status_code in [
-                job_models.STATUS_CODE_FAILED, job_models.STATUS_CODE_CANCELED]:
-            cls._add_error(
-                base_model_validators.ERROR_CATEGORY_ERROR_CHECK,
-                'Entity id %s: error for job is empty but '
-                'job status is %s' % (item.id, item.status_code))
-
-    @classmethod
-    def _validate_output(cls, item):
-        """Validate output for entity is present only if status is
-        completed.
-
-        Args:
-            item: ndb.Model. JobModel to validate.
-        """
-        if item.output and item.status_code != job_models.STATUS_CODE_COMPLETED:
-            cls._add_error(
-                base_model_validators.ERROR_CATEGORY_OUTPUT_CHECK,
-                'Entity id %s: output: %s for job is not empty but '
-                'job status is %s' % (item.id, item.output, item.status_code))
-
-        if item.output is None and (
-                item.status_code == job_models.STATUS_CODE_COMPLETED):
-            cls._add_error(
-                base_model_validators.ERROR_CATEGORY_OUTPUT_CHECK,
-                'Entity id %s: output for job is empty but '
-                'job status is %s' % (item.id, item.status_code))
-
-    @classmethod
-    def _get_custom_validation_functions(cls):
-        return [
-            cls._validate_time_fields,
-            cls._validate_error,
-            cls._validate_output]
 
 
 class ContinuousComputationModelValidator(
